Converted discounts to 4.0.
