--- ShopifySharp.Tests/ShopifyDiscountService Tests/Utils.cs
@@ -1,7 +1,0 @@
-﻿namespace ShopifySharp.Tests.ShopifyDiscountService_Tests
-{
-    static class DiscountUtils
-    {
-        public static ShopifyDiscountService Service = new ShopifyDiscountService(Utils.MyShopifyUrl, Utils.AccessToken);
-    }
-}
