+++ ShopifySharp/Entities/DiscountCode.cs
@@ -1,10 +1,4 @@
 using Newtonsoft.Json;
-using ShopifySharp.Enums;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ShopifySharp
 {
